Data.Text.Prettyprint.Doc.Minicute

Description

Additional prettyprinter functions for miniCUTE compiler

Documentation

class PrettyMC a where

PrettyMC (which stands for Pretty MiniCute) class is a variation class of Pretty for miniCUTE types, which require a precedence to print prettily.

Minimal complete definition

prettyMC

Methods

prettyMC :: Int -> a -> Doc ann

prettyListMC :: Int -> [a] -> Doc ann

prettyMC0 :: PrettyMC a => a -> Doc ann

prettyListMC0 :: PrettyMC a => [a] -> Doc ann

prettyIndent :: Doc ann -> Doc ann

prettyIndent doc make a document indented with an appropriate size.

prettyWrappedIf :: Bool -> (Doc ann -> Doc ann) -> Doc ann -> Doc ann

makePrettyMCFromPretty :: Name -> DecsQ